refactor: removed InsertRows as it is redundant after engine split
