+++ go/vt/vtgate/engine/insert_common.go
@@ -19,11 +19,14 @@ package engine
 import (
 	"context"
 	"fmt"
+	"strconv"
 
 	"vitess.io/vitess/go/sqltypes"
 	"vitess.io/vitess/go/vt/key"
 	querypb "vitess.io/vitess/go/vt/proto/query"
+	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/vterrors"
+	"vitess.io/vitess/go/vt/vtgate/evalengine"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
 )
 
@@ -58,6 +61,9 @@ type InsertCommon struct {
 	ForceNonStreaming bool
 
 	PreventAutoCommit bool
+
+	// Generate is only set for inserts where a sequence must be generated.
+	Generate *Generate
 
 	// ColVindexes are the vindexes that will use the VindexValues
 	ColVindexes []*vindexes.ColumnVindex
@@ -229,4 +235,143 @@ func (ic *InsertCommon) processUnowned(ctx context.Context, vcursor VCursor, vin
 	}
 
 	return nil
+}
+
+// processGenerateFromSelect generates new values using a sequence if necessary.
+// If no value was generated, it returns 0. Values are generated only
+// for cases where none are supplied.
+func (ic *InsertCommon) processGenerateFromSelect(
+	ctx context.Context,
+	vcursor VCursor,
+	rows []sqltypes.Row,
+) (insertID int64, err error) {
+	if ic.Generate == nil {
+		return 0, nil
+	}
+	var count int64
+	offset := ic.Generate.Offset
+	genColPresent := offset < len(rows[0])
+	if genColPresent {
+		for _, row := range rows {
+			if shouldGenerate(row[offset], evalengine.ParseSQLMode(vcursor.SQLMode())) {
+				count++
+			}
+		}
+	} else {
+		count = int64(len(rows))
+	}
+
+	if count == 0 {
+		return 0, nil
+	}
+
+	insertID, err = ic.execGenerate(ctx, vcursor, count)
+	if err != nil {
+		return 0, err
+	}
+
+	used := insertID
+	for idx, val := range rows {
+		if genColPresent {
+			if shouldGenerate(val[offset], evalengine.ParseSQLMode(vcursor.SQLMode())) {
+				val[offset] = sqltypes.NewInt64(used)
+				used++
+			}
+		} else {
+			rows[idx] = append(val, sqltypes.NewInt64(used))
+			used++
+		}
+	}
+
+	return insertID, nil
+}
+
+// processGenerateFromValues generates new values using a sequence if necessary.
+// If no value was generated, it returns 0. Values are generated only
+// for cases where none are supplied.
+func (ic *InsertCommon) processGenerateFromValues(
+	ctx context.Context,
+	vcursor VCursor,
+	bindVars map[string]*querypb.BindVariable,
+) (insertID int64, err error) {
+	if ic.Generate == nil {
+		return 0, nil
+	}
+
+	// Scan input values to compute the number of values to generate, and
+	// keep track of where they should be filled.
+	env := evalengine.NewExpressionEnv(ctx, bindVars, vcursor)
+	resolved, err := env.Evaluate(ic.Generate.Values)
+	if err != nil {
+		return 0, err
+	}
+	count := int64(0)
+	values := resolved.TupleValues()
+	for _, val := range values {
+		if shouldGenerate(val, evalengine.ParseSQLMode(vcursor.SQLMode())) {
+			count++
+		}
+	}
+
+	// If generation is needed, generate the requested number of values (as one call).
+	if count != 0 {
+		insertID, err = ic.execGenerate(ctx, vcursor, count)
+		if err != nil {
+			return 0, err
+		}
+	}
+
+	// Fill the holes where no value was supplied.
+	cur := insertID
+	for i, v := range values {
+		if shouldGenerate(v, evalengine.ParseSQLMode(vcursor.SQLMode())) {
+			bindVars[SeqVarName+strconv.Itoa(i)] = sqltypes.Int64BindVariable(cur)
+			cur++
+		} else {
+			bindVars[SeqVarName+strconv.Itoa(i)] = sqltypes.ValueBindVariable(v)
+		}
+	}
+	return insertID, nil
+}
+
+const nextValBV = "n"
+
+func (ic *InsertCommon) execGenerate(ctx context.Context, vcursor VCursor, count int64) (int64, error) {
+	// If generation is needed, generate the requested number of values (as one call).
+	rss, _, err := vcursor.ResolveDestinations(ctx, ic.Generate.Keyspace.Name, nil, []key.Destination{key.DestinationAnyShard{}})
+	if err != nil {
+		return 0, err
+	}
+	if len(rss) != 1 {
+		return 0, vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "auto sequence generation can happen through single shard only, it is getting routed to %d shards", len(rss))
+	}
+	bindVars := map[string]*querypb.BindVariable{nextValBV: sqltypes.Int64BindVariable(count)}
+	qr, err := vcursor.ExecuteStandalone(ctx, nil, ic.Generate.Query, bindVars, rss[0])
+	if err != nil {
+		return 0, err
+	}
+	// If no rows are returned, it's an internal error, and the code
+	// must panic, which will be caught and reported.
+	return qr.Rows[0][0].ToCastInt64()
+}
+
+// shouldGenerate determines if a sequence value should be generated for a given value
+func shouldGenerate(v sqltypes.Value, sqlmode evalengine.SQLMode) bool {
+	if v.IsNull() {
+		return true
+	}
+
+	// Unless the NO_AUTO_VALUE_ON_ZERO sql mode is active in mysql, it also
+	// treats 0 as a value that should generate a new sequence.
+	value, err := evalengine.CoerceTo(v, sqltypes.Uint64, sqlmode)
+	if err != nil {
+		return false
+	}
+
+	id, err := value.ToCastUint64()
+	if err != nil {
+		return false
+	}
+
+	return id == 0
 }
